+++ lib/workato/cli/oauth2_command.rb
@@ -101,6 +101,7 @@ module Workato
         return unless @thread
 
         Rack::Handler::WEBrick.shutdown
+        @thread.join
         @thread.exit
       end