+++ lib/workato/connector/sdk/blank.rb
@@ -1,0 +1,4 @@
+# typed: false
+# frozen_string_literal: true
+#
+# This file is intentionally left blank. Do not remove